+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Inline the removal of orphaned subscriptions as part of the response processing pipeline.
+
 ### [v6.137](https://github.com/replicant4j/replicant/tree/v6.137) (2024-11-15) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.136...v6.137)
 
 Changes in this release: